fx quant: move qconfig utils to utils file
Summary:

Improving readability

Test Plan:

CI

Reviewers:

Subscribers:

Tasks:

Tags:

[ghstack-poisoned]
